% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDSannotation_snv.R
\name{CDSannotation_snv}
\alias{CDSannotation_snv}
\title{snvRF-Annotations
Avoid tabix file error by placing the file in the same directory as your vcf
Tabix (tbi) is required for annotations or you will get this error:
"Error in (function (classes, fdef, mtable)  :
unable to find an inherited method for function âscanVcfâ for signature â"character", "GRanges"â"
Step 3 Annotations}
\usage{
CDSannotation_snv(x, predicted)
}
\description{
snvRF-Annotations
Avoid tabix file error by placing the file in the same directory as your vcf
Tabix (tbi) is required for annotations or you will get this error:
"Error in (function (classes, fdef, mtable)  :
unable to find an inherited method for function âscanVcfâ for signature â"character", "GRanges"â"
Step 3 Annotations
}
\examples{


}
