% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDSannotation_snv_whitelist.R
\name{CDSannotation_snv_whitelist}
\alias{CDSannotation_snv_whitelist}
\title{snvRF-Annotations
Step 3 Annotations
FOR WHITELIST ATTEMPT ANNOTATIONS IF POSSIBLE BUT MAY HAVE INACCURACIES}
\usage{
CDSannotation_snv_whitelist(x, predicted, build, change.build)
}
\arguments{
\item{x}{List object containing the four vcf.gz files from 
callers MuTect2, Freebayes, VarDict and VarScan.}

\item{predicted}{List object containing 1.stats, 2.predicted-snv and 3.parse-snv matrices}

\item{build}{Your current genomic build}

\item{change.build}{Choose TRUE to convert from 'hg19.to.hg38' or 'hg38.to.hg19' annotation output}
}
\description{
snvRF-Annotations
Step 3 Annotations
FOR WHITELIST ATTEMPT ANNOTATIONS IF POSSIBLE BUT MAY HAVE INACCURACIES
}
\examples{


}
