% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.glmsmurf}
\alias{coef.glmsmurf}
\alias{coefficients.glmsmurf}
\title{Coefficients of Estimated Model}
\usage{
\method{coef}{glmsmurf}(object, ...)

\method{coefficients}{glmsmurf}(object, ...)
}
\arguments{
\item{object}{An object of class '\code{\link[=glmsmurf-class]{glmsmurf}}', typically the result of a call to \code{\link{glmsmurf}} or \code{\link{glmsmurf.fit}}.}

\item{...}{Additional arguments which are currently ignored.}
}
\value{
A vector containing the coefficients of the estimated model in \code{object}.
}
\description{
Function to extract the coefficients of the estimated model. 
             \code{coefficients} is an \emph{alias} for it.
}
\examples{
## See example(glmsmurf) for examples

}
\seealso{
\code{\link{coef_reest}}, \code{\link[stats]{coef}}, \code{\link{summary.glmsmurf}}, \code{\link{glmsmurf}}, \code{\link{glmsmurf-class}}
}
