% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance.glmsmurf}
\alias{deviance.glmsmurf}
\title{Deviance of Estimated Model}
\usage{
\method{deviance}{glmsmurf}(object, ...)
}
\arguments{
\item{object}{An object of class '\code{\link[=glmsmurf-class]{glmsmurf}}', typically the result of a call to \code{\link{glmsmurf}} or \code{\link{glmsmurf.fit}}.}

\item{...}{Additional arguments which are currently ignored.}
}
\value{
The deviance of the estimated model in \code{object}.
}
\description{
Function to extract the deviance of the estimated model.
}
\examples{
## See example(glmsmurf) for examples
         
}
\seealso{
\code{\link{deviance_reest}}, \code{\link[stats]{deviance}}, \code{\link{summary.glmsmurf}}, 
         \code{\link{glmsmurf}}, \code{\link{glmsmurf-class}}
}
