% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance_reest}
\alias{deviance_reest}
\alias{deviance_reest.glmsmurf}
\title{Deviance of Re-estimated Model}
\usage{
deviance_reest(object, ...)

\method{deviance_reest}{glmsmurf}(object, ...)
}
\arguments{
\item{object}{An object for which the extraction of the deviance is meaningful. 
E.g. an object of class '\code{\link[=glmsmurf-class]{glmsmurf}}', typically the result of a call to \code{\link{glmsmurf}} or \code{\link{glmsmurf.fit}}.}

\item{...}{Additional arguments which are currently ignored.}
}
\value{
The deviance of the re-estimated model in \code{object}, 
        when it is available or, otherwise, the deviance of the estimated model in \code{object} with a warning.
}
\description{
Function to extract the deviance of the re-estimated model.
}
\examples{
## See example(glmsmurf) for examples
   
}
\seealso{
\code{\link{deviance.glmsmurf}}, \code{\link[stats]{deviance}}, \code{\link{summary.glmsmurf}}, 
         \code{\link{glmsmurf}}, \code{\link{glmsmurf-class}}
}
