% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlag.R
\name{findlag}
\alias{findlag}
\title{Create lag variables}
\usage{
findlag(date, x, af_lags = 1, be_lags = 0, by)
}
\arguments{
\item{date}{A vector of dates.}

\item{x}{The variable to be lagged.}

\item{af_lags}{The number of days following event days.}

\item{be_lags}{The number of days before event days.}

\item{by}{A vector of ids or a matrix with columns as the id variables. The events will be found 
separately within each unique combination of id variables. This is optional.}
}
\value{
A \code{data.table} object with lagged exposures as well as the date and by variables.
}
\description{
This function finds non-event days that are near events (e.g. the days before or after heat waves or 
ozone events). The days are potentially excluded as potential control days when using matching 
approachs.
}
\author{
Ander Wilson
}
