% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.glmsmurf}
\alias{fitted.glmsmurf}
\title{Fitted Values of Estimated Model}
\usage{
\method{fitted}{glmsmurf}(object, ...)
}
\arguments{
\item{object}{An object of class '\code{\link[=glmsmurf-class]{glmsmurf}}', typically the result of a call to \code{\link{glmsmurf}} or \code{\link{glmsmurf.fit}}.}

\item{...}{Additional arguments which are currently ignored.}
}
\value{
A vector containing the fitted values of the estimated model in \code{object}.
}
\description{
Function to extract the fitted values of the estimated model.
}
\examples{
## See example(glmsmurf) for examples
    
}
\seealso{
\code{\link{fitted_reest}}, \code{\link[stats:fitted.values]{fitted}}, \code{\link{glmsmurf}}, \code{\link{glmsmurf-class}}
}
