% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted_reest}
\alias{fitted_reest}
\alias{fitted_reest.glmsmurf}
\title{Fitted Values of Re-estimated Model}
\usage{
fitted_reest(object, ...)

\method{fitted_reest}{glmsmurf}(object, ...)
}
\arguments{
\item{object}{An object for which the extraction of fitted values is meaningful. 
E.g. an object of class '\code{\link[=glmsmurf-class]{glmsmurf}}', typically the result of a call to \code{\link{glmsmurf}} or \code{\link{glmsmurf.fit}}.}

\item{...}{Additional arguments which are currently ignored.}
}
\value{
A vector containing the fitted values of the re-estimated model in \code{object}, 
        when they are available or, otherwise, the fitted values of the estimated model in \code{object} with a warning.
}
\description{
Function to extract the fitted values of the re-estimated model.
}
\examples{
## See example(glmsmurf) for examples
 
}
\seealso{
\code{\link{fitted.glmsmurf}}, \code{\link[stats:fitted.values]{fitted}}, \code{\link{glmsmurf}}, \code{\link{glmsmurf-class}}
}
