% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpd.R
\name{hpd}
\alias{hpd}
\title{Calculate highest posterior density intervals for MCMC sample}
\usage{
hpd(x, prob = 0.95)
}
\arguments{
\item{x}{A numeric vector to summarize.}

\item{prob}{The probability that the interval will contain. The default is 0.95.}

\item{...}{Additional arguments to be passed to ...}
}
\value{
The endpoints of the interval
}
\description{
Find the highest posterior density intervals for a vector x.
}
\author{
Ander Wilson
}
