% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indelRFpredict.R
\name{indelRFpredict}
\alias{indelRFpredict}
\title{indelRF-Predict}
\usage{
indelRFpredict(parsevcf, indel.cutoff)
}
\description{
Indel prediction model SMuRFv1.5
Step 2 Predict
}
\examples{


}
