% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{lag}
\alias{lag}
\title{Create lag variables}
\usage{
lag(date, x, lags = 1, by)
}
\arguments{
\item{date}{A vector of dates.}

\item{x}{The varibale to be lagged}

\item{lags}{The number of days the variable should be lagged}

\item{by}{A vector of ids or a matrix with columns as the id variables. The events will be found 
separately within each unique combination of id variables. This is optional.}
}
\value{
A \code{data.table} object with lagged exposures as well as the date and by variables.
}
\description{
This function finds non-event days that are near events (e.g. the day after a heat waves or 
ozone events). The days are potentially excluded as potential control days when using matching 
approachs.
}
\author{
Ander Wilson
}
