% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtch.R
\name{mtch}
\alias{mtch}
\title{Match}
\usage{
mtch(date, casecontrol, matchvars = NULL, mahdoy = FALSE, caldays = Inf,
  by = NULL, ratio = 1, seed = NULL, ...)
}
\arguments{
\item{date}{A vector of dates.}

\item{casecontrol}{A vector where 1 indentifies events, 0 identifies potential controls, and 
\code{NA} represent non-event days that are ineligible as matches (e.g. because of missing data, 
or they are too close to other event days, as determined by \code{findeventneighbors}).}

\item{matchvars}{A matrix where the columns represent variables that will be matched on using 
the Mahalanobis distance and nearest neighbor.}

\item{mahdoy}{A logical indicating if the day of year should be included in the 
Mahalanobis-metric matching within the caliper. The default is FALSE. If TRUE then the 
distance to the treatment day based on day of year will be included with the other matching 
variables.}

\item{caldays}{The number of days that matched value will be selected from. This ignores year 
but only looks at date and month. For example, if \code{caldays} is set to 7 then the control 
day will be from within 7 days of the day of the year of the event day (inclusive).}

\item{by}{A vector of IDs or a matrix with columns as the ID variables. The events will be 
found separately within each unique combination of ID variables. For example, if the dataset
includes multiple cities, `by` could be used to ensure that matching is always done within 
city.}

\item{ratio}{The number of control days to be matched to each event day. The default is 1.  See 
documentation for \code{matchit} for more details.}

\item{seed}{A seed for a random number generator.}

\item{...}{Additional arguments to be passed to \code{matchit}}
}
\value{
data A \code{data.table} object of event days and their matched controls.

nn A summary of the number of event days and controls in the final matched dataset. 
   See documentation for \code{matchit} for more details.

sum.matched A summary of the quality of each match. See documentation for 
   \code{matchit} for more details.
}
\description{
Finds matches for the events without a strata and within a time window. The Mahalanobis-metric 
can be used to match on other variables as well. This function is a wrapper for the 
\code{matchit} function from the \code{MatchIt} package.
}
\seealso{
MatchIt
}
\author{
Ander Wilson
}
