% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcutoffs.R
\name{newcutoffs}
\alias{newcutoffs}
\title{newcutoffs}
\usage{
newcutoffs(parse.dir = NULL, snv.cutoff = "default",
  indel.cutoff = "default")
}
\arguments{
\item{parse.dir}{Directory to parse files}

\item{snv.cutoff}{}

\item{indel.cutoff}{}

\item{output.dir}{your output directory}
}
\description{
newcutoffs
}
\examples{


}
