% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsevcf.R
\name{parsevcf}
\alias{parsevcf}
\title{parsevcf}
\usage{
parsevcf(x)
}
\arguments{
\item{filedir}{Sets the working directory containing the vcf.gz files from 
callers MuTect2, Freebayes, VarDict and VarScan.  
include Allele frequency here}
}
\description{
Initial file parse and preparation before filtering
}
\examples{


}
