% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsevcf_raw.R
\name{parsevcf_raw}
\alias{parsevcf_raw}
\title{parsevcf}
\usage{
parsevcf_raw(x)
}
\arguments{
\item{filedir}{Sets the working directory containing the vcf.gz files from 
callers MuTect2, Freebayes, VarDict and VarScan.  
include Allele frequency here}
}
\description{
Step 1 Initial parse and filtering
}
\examples{


}
