% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.glmsmurf}
\alias{plot.glmsmurf}
\title{Plot Coefficients of Estimated Model}
\usage{
\method{plot}{glmsmurf}(x, xlab = "Index", ylab = "Estimated coefficients", basic = FALSE, ...)
}
\arguments{
\item{x}{An object of class '\code{\link[=glmsmurf-class]{glmsmurf}}', typically the result of a call to \code{\link{glmsmurf}} or \code{\link{glmsmurf.fit}}.}

\item{xlab}{Label for the x-axis, default is \code{"Index"}.}

\item{ylab}{Label for the y-axis, default is \code{"Estimated coefficients"}.}

\item{basic}{Logical indicating if the basic lay-out is used for the plot, default is \code{FALSE}.}

\item{...}{Additional arguments for the \code{\link[graphics:plot.default]{plot}} function.}
}
\description{
Function to plot the coefficients of the estimated model.
}
\details{
When \code{basic=FALSE}, the improved lay-out for the plot is used. Per predictor, groups of equal coefficients are indicated
in the same color (up to 8 colors), and zero coefficients are indicated by grey squares.
}
\examples{
## See example(glmsmurf) for examples
}
\seealso{
\code{\link{plot_reest}}, \code{\link{coef.glmsmurf}}, \code{\link{summary.glmsmurf}}, \code{\link{glmsmurf}}, \code{\link{glmsmurf-class}}
}
