% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.vcf.R
\name{read.vcf}
\alias{read.vcf}
\title{parsevcf}
\usage{
read.vcf(x, roi = F, roi.dir = NULL, t.label)
}
\arguments{
\item{x}{List object containing the four vcf.gz files from 
callers MuTect2, Freebayes, VarDict and VarScan.}

\item{roi}{set to TRUE to enable this function. Alternatively read.vcf is a VCF parser in development.}

\item{roi.dir}{Directory to BED file of genomic coordinates.}

\item{t.label}{Tumor sample name label}
}
\description{
Step 1 Initial parse and filtering
}
\examples{


}
