% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.files.R
\name{save.files}
\alias{save.files}
\title{save.files}
\usage{
save.files(myresults, output.dir)
}
\arguments{
\item{myresults}{List object for SMuRF output}

\item{output.dir}{your output directory}
}
\description{
save.files
}
