% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smurf.R
\docType{package}
\name{smurf-package}
\alias{smurf}
\alias{smurf-package}
\title{smurf: Sparse Multi-Type Regularized Feature Modeling}
\description{
Implementation of the SMuRF algorithm of Devriendt et al. (2018) <arXiv:1810.03136> to fit generalized linear models (GLMs) with multiple types of predictors via regularized maximum likelihood.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/TReynkens/smurf}
  \item Report bugs at \url{https://gitlab.com/TReynkens/smurf/-/issues}
}

}
\author{
\strong{Maintainer}: Tom Reynkens \email{tomreynkens@hotmail.com} (\href{https://orcid.org/0000-0002-5516-5107}{ORCID})

Authors:
\itemize{
  \item Sander Devriendt \email{sander.devriendt@kuleuven.be}
  \item Katrien Antonio
}

}
