% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.glmsmurf}
\alias{summary.glmsmurf}
\title{Summary of a Multi-Type Regularized GLM Fitted Using the SMuRF Algorithm}
\usage{
\method{summary}{glmsmurf}(object, digits = 3L, ...)
}
\arguments{
\item{object}{An object of class '\code{\link[=glmsmurf-class]{glmsmurf}}', typically the result of a call to \code{\link{glmsmurf}} or \code{\link{glmsmurf.fit}}.}

\item{digits}{The number of significant digits used when printing, default is 3.}

\item{...}{Additional arguments which are currently ignored.}
}
\description{
Function to print a summary of a \code{glmsmurf}-object.
}
\examples{
## See example(glmsmurf) for examples

}
\seealso{
\code{\link[stats]{summary.glm}}, \code{\link{glmsmurf}}, \code{\link{glmsmurf-class}}
}
