% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDSannotation_snv.R
\name{CDSannotation_snv}
\alias{CDSannotation_snv}
\title{snvRF-Annotations
Step 3 Annotations}
\usage{
CDSannotation_snv(x, predicted, build, change.build)
}
\arguments{
\item{x}{List object containing the four vcf.gz files from 
callers MuTect2, Freebayes, VarDict and VarScan.}

\item{predicted}{List object containing 1.stats, 2.predicted-snv and 3.parse-snv matrices}

\item{build}{Your current genomic build}

\item{change.build}{Choose TRUE to convert from 'hg19.to.hg38' or 'hg38.to.hg19' annotation output}
}
\description{
snvRF-Annotations
Step 3 Annotations
}
\examples{


}
