% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findeventneighbors.R
\name{findeventneighbors}
\alias{findeventneighbors}
\title{Find event neighbors}
\usage{
findeventneighbors(date, event, days = 0, by)
}
\arguments{
\item{date}{A vector of dates.}

\item{event}{A logical indicating if a day was an event.}

\item{days}{The number of days that must separate a given day from an event.}

\item{by}{A vector of ids or a matrix with columns as the id variables. The events will be found separately within each unique combination of id variables. This is optional.}
}
\value{
Returns a vector of logicals that indivates which days are not events and not with in days (number provided) of another event.
}
\description{
This function finds non-event days that are near events (e.g. the day after a heat waves or ozone events). The days are potentially excluded as potential control days when using matching approachs.
}
\author{
Ander Wilson
}

