% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findevents.R
\name{findevents}
\alias{findevents}
\title{Find events}
\usage{
findevents(date, x, xmin, mindays = 2, by)
}
\arguments{
\item{date}{A vector of dates.}

\item{x}{The exposure variable (e.g. temperature for heat waves or ozone for ozone events)}

\item{xmin}{The minimum value that x must be for at least mindays to qualify as an event.}

\item{mindays}{The minimum number of consecutive days that qualify as an event.}

\item{by}{A vector of ids or a matrix with columns as the id variables. The events will be 
found separately within each unique combination of id variables. This is optional.}
}
\value{
Returns a data.table with the columns for by, x, date as well as the following new 
   variables: above and indivator the x>xmin; above_fromstart number of consecutive days that 
   have exceeded xmin through the current date; above_toend the number of remaining days that 
   are above xmin; length the total number of consecutive days that are above xmin; event a 
   logical indivatring a day is part of an event.
}
\description{
This function finds events (e.g. heat waves or ozone events). The approach is to find 
consecutive days that are above a thresholds.
}
\author{
Ander Wilson
}
