% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hpd.R
\name{hpd}
\alias{hpd}
\title{Calculate highest posterior density intervals for MCMC sample}
\usage{
hpd(x, prob = 0.95)
}
\arguments{
\item{x}{A numeric vector to summarize.}

\item{prop}{The probability that the interval will contain. The default is 0.95.}
}
\value{
The endpoints of the interval
}
\description{
this function finds the highest posterior density intervals for a vector x.
}
\author{
Ander Wilson
}

