% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indelRFpredict.R
\name{indelRFpredict}
\alias{indelRFpredict}
\title{indelRF-Predict}
\usage{
indelRFpredict(parsevcf, indel.cutoff, fixed.indel.cutoff = F)
}
\arguments{
\item{parsevcf}{List object containing 1.snv-parse and 2.indel-parse matrices.}

\item{indel.cutoff}{Threshold for true predictions by SMuRF}

\item{fixed.indel.cutoff}{debug for when threshold is zero.}
}
\description{
Indel prediction model SMuRFv2.0 (>R-3.5.1)
Step 2 Predict
}
\examples{


}
