% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackse.R
\name{jackse}
\alias{jackse}
\title{Jackknife standard errors}
\usage{
jackse(x, theta, bycol = TRUE, ...)
}
\arguments{
\item{x}{A vector or matrix to be summarized.}

\item{theta}{The summary statistic to be used. For example, \code{theta = mean} calculates the 
jackknife standard errors for the mean.}

\item{bycol}{Logical indicating if the data are in rows or columns when x is a matrix. The 
default is TRUE.}

\item{...}{Other parameters to be passed to the summary statistic function specified by 
\code{theta}.}
}
\value{
Jackknife standard errors.
}
\description{
Calculate jackknife standard errors.
}
\author{
Ander Wilson
}
