% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hpd copy.R
\name{jackse}
\alias{jackse}
\title{Jackknife standard errors}
\usage{
jackse(x, theta, bycol = TRUE, ...)
}
\arguments{
\item{x}{A vector or matrix to be summarized.}

\item{theta}{The summary statistic to be used. For example, theta=mean calcualted the jackknife standard errors for the mean.}

\item{bycol}{Logical indicating if the data are in rows or columns when x is a matrix. The default is TRUE.}

\item{...}{Other parameters to be passed to the function theta().}
}
\value{
Jackknife standard errors.
}
\description{
This function calcualte jackknife standard errors
}
\author{
Ander Wilson
}

