% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matchmahal.R
\name{matchmahal}
\alias{matchmahal}
\title{Match with Mahalanobis}
\usage{
matchmahal(date, casecontrol, matchvars, by)
}
\arguments{
\item{date}{A vector of dates.}

\item{casecontrol}{A vector where 1 indentifies events, 0 identifies potential controls, and NA represent non-event days that are not eligible to be matched (e.g. missing data, or excluded for some other reason).}

\item{matchvars}{A matrix where the columns represent variables that will be matched on using the Mahalanobis distrance and nearest neighbor.}

\item{by}{A vector of ids or a matrix with columns as the id variables. The events will be found separately within each unique combination of id variables. This is optional.}
}
\value{
data A data.table of matched cases and controls.

nn A summary of the number if cases and controls that were matched. See documentation for matchit for more details.

sum.matched A summary of the quality of each match. See documentation for matchit for more details.
}
\description{
This function finds matches for the events without a strata sing the Mahalanobis distrance and nearest neighbor. It is a wrapper for the matchit function from the MatchIt package.
}
\author{
Ander Wilson
}

