% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsevcf_allfeaturesall.R
\name{parsevcf_allfeaturesall}
\alias{parsevcf_allfeaturesall}
\title{parsevcf}
\usage{
parsevcf_allfeaturesall(x, roi = F, roi.dir = NULL, t.label = NULL)
}
\arguments{
\item{x}{List object containing the four vcf.gz files from 
callers MuTect2, Freebayes, VarDict and VarScan.}

\item{t.label}{Default='-T'. Identify your tumour sample label correctly.}
}
\description{
Step 1 Initial parse and filtering
}
\examples{


}
