% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smurf.R
\name{smurf}
\alias{smurf}
\title{SMuRF v1.5}
\usage{
smurf(directory = NULL, model, nthreads = -1, snv.cutoff = "default",
  indel.cutoff = "default")
}
\arguments{
\item{directory}{Choose directory where the files Variant Caller Format(VCF) files are located.}

\item{model}{Choose either "combined" or "cdsannotation"
The appropriate parsing and prediction model will be performed
to obtain a list of somatic mutation calls. 
Choose "combined" to generate both SNV and Indel outputs
Choose "cdsannotation" to run "combined" + add annotations to coding regions (from the respective coding transcripts)}

\item{nthreads}{Default as "-1", where all available cores will be used for RandomForest prediction. 
Specify any integer from 1 to x, depending on your resources available.}

\item{snv.cutoff}{Default SNV model cutoff, unless a number between 0 to 1 is stated.}

\item{indel.cutoff}{Default indel model cutoff, unless a number between 0 to 1 is stated.}
}
\description{
Somatic mutation consensus calling based on four callers:
MuTect2, Freebayes, VarDict, VarScan
using a RandomForest model to consolidate a list of high accuracy calls.
}
\note{
Input files containing variant calls should be ".vcf.gz" format of each caller.
Supported for R (>=3.3.1)
}
\examples{
smurf("/path/to/directory..","combined")
smurf("/path/to/directory..","cdsannotation")

}
