% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smurf.R
\name{smurf}
\alias{smurf}
\title{SMuRF v1.6}
\usage{
smurf(directory = NULL, mode = NULL, nthreads = -1, annotation = F,
  output.dir = NULL, parse.dir = NULL, whitelist.file = NULL,
  snv.cutoff = "default", indel.cutoff = "default", build = NULL,
  change.build = F, check.packages = T)
}
\arguments{
\item{directory}{Choose directory where the files Variant Caller Format(VCF) files are located. 
Alternatively, provide a list object containing the path to the 4 VCF files labelled: mutect, freebayes, vardict and varscan.}

\item{mode}{Choose "snv", "indel" or "combined" (snv+indel). 
The appropriate parsing and prediction model will be performed
to obtain a list of somatic mutation calls. 

Debug mode: Choose "featureselectionall" to parse the VCF files only.}

\item{nthreads}{Default as "-1", where all available cores will be used for RandomForest prediction. 
Specify any integer from 1 to x, depending on your resources available.
For 32-bit Windows, only 1 core is allowed (nthreads=1).}

\item{annotation}{TRUE or FALSE (default). Provide gene annotations for each variant call.}

\item{output.dir}{Path to output directory (if saving files as .txt)}

\item{parse.dir}{Specify if changing SMuRF default cutoffs.
Path to the location of existing snv-parse.txt and indel-parse.txt files generated by SMuRF.}

\item{whitelist.file}{Path to .BED or .txt file with region to be parsed into the SMuRF format.
All SNVs and INDELS will be retrieved. SMuRF prediction will not be performed.}

\item{snv.cutoff}{Default SNV model cutoff, unless a number between 0 to 1 is stated.}

\item{indel.cutoff}{Default indel model cutoff, unless a number between 0 to 1 is stated.}

\item{build}{Genome build='hg19' or "hg38". If unknown, SMuRF will try to detect it.}

\item{change.build}{TRUE or FALSE. For conversion of your genomic coordinates.
Default option is disabled to retain your original 'build' specified above.}

\item{check.packages}{Default as TRUE. For debug mode.}
}
\description{
Somatic mutation consensus calling based on four callers:
MuTect2, Freebayes, VarDict, VarScan
using a RandomForest model to consolidate a list of high accuracy calls.
}
\note{
Input files containing variant calls should be ".vcf.gz" format of each caller.
Supported for R (>=3.3.1), Java 7 or 8. For R (>=3.5.1), Java (7 up to 11 is supported).
}
\examples{
myresults = smurf(directory="/path/to/directory..",
                  mode="indel")

myresults = smurf(directory="/path/to/directory..",
                  mode="snv",
                  save.files=T, output.dir="/path/to/output")

myresults = smurf(directory="/path/to/directory..",
                  mode="combined",
                  nthreads = 4,
                  snv.cutoff = 0.2, indel.cutoff = 'default',
                  parse.dir="/path/to/parse/files/")

myresults = smurf(directory="/path/to/directory..",
                  mode="combined", 
                  annotation=T, build='hg19')
                  
myresults = smurf(directory="/path/to/directory..",
                  whitelist.dir="/path/to/dir/roi.bed")

#Change hg38 to hg19 coordinates in gene annotation output
myresults = smurf(directory="/path/to/directory..",
                  mode="combined", 
                  annotation=T, 
                  build='hg38', change.build=T)
                  
#Specify directories manually
dir.list = list(mutect='/path/to/mutect.vcf.gz',
                freebayes='/path/to/freebayes.vcf.gz',
                vardict='/path/to/vardict.vcf.gz',
                varscan='/path/to/varscan.vcf.gz')
                
myresults = smurf(directory=dir.list, 
                  model="combined")
                                   
}
