% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smurf.R
\name{smurf}
\alias{smurf}
\title{SMuRF v1.3}
\usage{
smurf(directory, model, nthreads = -1, ncores = 1)
}
\arguments{
\item{directory}{Choose directory where the files Variant Caller Format(VCF) files are located.}

\item{model}{Choose either "snv" or "indel" or "combined"
The appropriate parsing and prediction model will be performed
to obtain a list of somatic mutation calls. 
Choose "combined" to generate both SNV and Indel outputs}
}
\description{
Somatic mutation consensus calling based on four callers:
MuTect2, Freebayes, VarDict, VarScan
using a RandomForest model to consolidate a list of high accuracy calls.
}
\note{
Input files containing variant calls should be ".vcf.gz" format of each caller.
Supported for R (>=3.3.1)
}
\examples{
smurf("/path/to/directory..","combined")
smurf("/path/to/directory..","cdsannotation")

}
