% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvRFpredict.R
\name{snvRFpredict}
\alias{snvRFpredict}
\title{snvRF-Predict}
\usage{
snvRFpredict(parsevcf, snv.cutoff, fixed.snv.cutoff = F)
}
\arguments{
\item{parsevcf}{List object containing 1.snv-parse and 2.indel-parse matrices.}

\item{snv.cutoff}{Threshold for true predictions by SMuRF}

\item{fixed.snv.cutoff}{debug for when threshold is zero.}
}
\description{
snv prediction model
Step 2 Predict
}
